requirements.txt
fastapi
asyncpg
python-dotenv
uvicorn
aiohttp
psycopg2-binary
sqlalchemy